Move the const casting code into its dedicated file
